Renamed notification component
